feat(animation): add animation getting started developer guide
BREAKING CHANGE: animation and transition mixins are now namespaced differently
