+++ guide/src/components/Drop.module.scss
@@ -32,8 +32,6 @@ $cardPadding: $ca-grid / 2;
 }
 
 .example {
-  animation-duration: 1s;
-  animation-fill-mode: both;
   background: #333;
   border-radius: 3px;
   height: 60px;
