+++ guide/src/pages/guides/animation/examples/AnimationImportExample.scss
@@ -1,0 +1,1 @@
+@import '~cultureamp-style-guide/styles/animations';
