+++ guide/src/pages/guides/animation/examples/AnimationUsageInputExample.js
@@ -1,0 +1,11 @@
+import scssCode from '!raw-loader!./AnimationUsageInputExample.scss';
+import React from 'react';
+import Code from '../../../../components/Code';
+
+class AnimationUsageInputExample extends React.PureComponent {
+  render() {
+    return <Code>{scssCode}</Code>;
+  }
+}
+
+export default AnimationUsageInputExample;
