+++ guide/src/pages/guides/animation/examples/AnimationUsageOutputExample.js
@@ -1,0 +1,11 @@
+import scssCode from '!raw-loader!./AnimationUsageOutputExample.scss';
+import React from 'react';
+import Code from '../../../../components/Code';
+
+class AnimationUsageOutputExample extends React.PureComponent {
+  render() {
+    return <Code>{scssCode}</Code>;
+  }
+}
+
+export default AnimationUsageOutputExample;
