+++ guide/src/pages/guides/animation/examples/TransitionUsageInputExample.js
@@ -1,0 +1,11 @@
+import scssCode from '!raw-loader!./TransitionUsageInputExample.scss';
+import React from 'react';
+import Code from '../../../../components/Code';
+
+class TransitionUsageInputExample extends React.PureComponent {
+  render() {
+    return <Code>{scssCode}</Code>;
+  }
+}
+
+export default TransitionUsageInputExample;
