+++ guide/src/pages/guides/animation/examples/TransitionUsageInputExample.scss
@@ -1,0 +1,13 @@
+//////////////////////////////////////////////
+// Fade In Transition
+//////////////////////////////////////////////
+.ca-animation-fade-in {
+  @include ca-animation-fade($state: in, $duration: 500ms, $delay: 250ms);
+}
+
+//////////////////////////////////////////////
+// Fade Out Transition
+//////////////////////////////////////////////
+.ca-animation-fade-out {
+  @include ca-animation-fade($state: out, $duration: 500ms, $delay: 250ms);
+}
