+++ guide/src/pages/index.md
@@ -7,7 +7,7 @@ imports:
 
 Product:
 
-* <Link to="/styles/">Foundational Styles</Link>
+* <Link to="/styles/">Foundation</Link>
 
   * <Link to="/styles/animation/">Animation</Link>
 
@@ -37,6 +37,8 @@ Product:
 
 * <Link to="/guides/">Guides</Link>
 
-  * <Link to="/language/">Product Language Style Guide</Link>
+  * <Link to="/guides/animation">Animation Guide</Link>
 
   * <Link to="/guides/localisation">Localisation Guide</Link>
+
+  * <Link to="/language/">Product Language Style Guide</Link>
