--- guide/src/pages/styles/animation/examples/_AnimationSequences.scss
@@ -1,16 +1,0 @@
-@import '~cultureamp-style-guide/styles/animation';
-
-$duration: $ca-duration-deliberate; // 700ms
-$delay: -$ca-duration-slow; // -400ms (stagger effect)
-
-@include ca-series {
-  .spin-1 {
-    @include ca-queue($duration, $delay, spin, fade($from: 0.75));
-  }
-  .spin-2 {
-    @include ca-queue($duration, $delay, spin, fade($from: 0.5));
-  }
-  .spin-3 {
-    @include ca-queue($duration, 0, spin, fade($from: 0.25));
-  }
-}
