+++ guide/src/pages/styles/index.md
@@ -3,7 +3,7 @@ imports:
   Link: components/Link.js
 ---
 
-# Foundational Styles
+# Foundation
 
 > Elements of visual design that we seek to use consistently.
