+++ styles/animations/_durations.scss
@@ -1,5 +1,3 @@
-// Timing variables
-// These timing values are designed to snap together. 100+300=400. 300+400=700 etc.
 $ca-duration-instant: 0ms;
 $ca-duration-immediate: 100ms;
 $ca-duration-rapid: 200ms;
