+++ styles/animations/_easings.scss
@@ -1,6 +1,3 @@
-///////////////////////////////////////////////
-// EASING PRESETS                            //
-///////////////////////////////////////////////
 $ca-ease-in-out: cubic-bezier(0.455, 0.03, 0.515, 0.955);
 $ca-ease-in: cubic-bezier(0.55, 0.085, 0.68, 0.53);
 $ca-ease-out: cubic-bezier(0.25, 0.46, 0.45, 0.94);
